Remove `fuzz/src/lib.rs` now that #1830 is merged
